checkpoint 12-May-2025 SY

+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -3,11 +3,14 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Category\StoreCategoryRequest;
+use App\Http\Requests\Category\CountCategoryRequest;
 use App\Models\Category;
+use App\Models\Product;
+use App\Models\SubCategory;
 use App\Services\CategoryService;
 use Exception;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Validator;
 use InvalidArgumentException;
 use Symfony\Component\HttpFoundation\Response;
@@ -15,43 +18,25 @@ use Symfony\Component\HttpFoundation\Response;
 class CategoryController extends Controller
 {
 
-
     public function __construct(
         protected CategoryService $categoryService
     ) {
    
     }
 
-    public function index()
+    public function counts(CountCategoryRequest $request)//si
     {
-        $categories = $this->categoryService->getAllCategories();
+  
+        $inventory_id = $validatedData['inventory_id'] ?? null;
+   
 
-        return response()->success([
-            'Categories' => $categories
-        ], Response::HTTP_OK);
-    }
-    
-    public function counts()
-    {
-        $inventory_id = request('inventory_id');
-        $product_id = request('product_id');
-
-        $categories = $this->categoryService->getCategoryCounts($inventory_id , $product_id);
+        $categories = $this->categoryService->getCategoryCounts($inventory_id , $request->validated('product_id'));
 
         return response()->success([
            $categories
         ], Response::HTTP_OK);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create(Request $request)
-    {
-        //
-    }
 
     /**
      *
@@ -60,30 +45,13 @@ class CategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreCategoryRequest $request)//si
     {
         try {
-            $section_id = request('section_id');
-            $validate = Validator::make(
-                $request->all(),
-                [
-                    'name_en' => 'required|string|max:255',
-                    'name_ar' => 'required|string|max:255',
-                    'image' => 'required|image|mimes:jpeg,bmp,png,webp,svg'
-                ]
-            );
 
-            if ($validate->fails()) {
-                return response()->error(
-                    $validate->errors()
-                    ,Response::HTTP_OK
-                );
-            }
-
-            $validated_data = $validate->validated();
-
-            $category = $this->categoryService->createCategory($validated_data, $section_id, $section_id);
-
+      
+            $category = $this->categoryService->createCategory($request->validated(), $request->validated('section_id'));
+            
             return response()->success(
                 $category
                 , Response::HTTP_OK);
@@ -94,43 +62,7 @@ class CategoryController extends Controller
             );
         }
     }
-    
-    
-    
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
-    public function show()
-    {
-        try {
-            $category_id = request('category_id');
-            $category = $this->categoryService->getCategory($category_id);
-
-            return response()->success(
-                $category
-                ,Response::HTTP_OK);
-        } catch (InvalidArgumentException $e) {
-            return response()->error(
-                $e->getMessage(),
-                Response::HTTP_NOT_FOUND
-            );
-        }
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Category $category)
-    {
-        //
-    }
 
     /**
      * Update the specified resource in storage.
@@ -139,43 +71,26 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request)
+    public function update(Request $request)//si
     {
         try {
             $section_id = request('section_id');
-            $category_id = request('category_id');
-
-            $validate = Validator::make(
-                $request->all(),
-                [
-                    'name_en' => 'required|string|max:255',
-                    'name_ar' => 'required|string|max:255',
-                    'image' => 'nullable|image|mimes:jpeg,bmp,png,webp,svg'
-                ]
-            );
-
-            if ($validate->fails()) {
-                return response()->error(
-                    $validate->errors(),
-                    Response::HTTP_OK
-                );
-            }
-
-            $validated_data = $validate->validated();
-
+            $category_id = request('category_id');		
             $has_photo = false;
+
             if ($request->hasFile('image')) {
                 $has_photo = true;
             }
 
-            $category = $this->categoryService->updateCategory($validated_data, $category_id, $section_id, $has_photo);
-
+            $category = $this->categoryService->updateCategory($request->all(), $category_id, $section_id, $has_photo);
+            
             return response()->success(
                 [
                     'data' => $category
                 ],
                 Response::HTTP_CREATED
             );
+
         } catch (InvalidArgumentException $e) {
             return response()->error(
                 $e->getMessage(),
@@ -190,10 +105,20 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function destroy()
+    public function destroy(Request $request)//si
     {
         try {
-            $category_id = request('category_id');
+            $validatedData = $request->validate([
+				'category_id' => ['required', 'integer', 'exists:categories,id'],
+			]);
+			$category_id = $validatedData['category_id'];
+			$category_product = Category::find($category_id)->products()->first();
+			if(isset($category_product)){
+				return response()->error(
+                'You Can NOT Remove This Category Cause It Has Products',
+                400
+            );	
+			}
             $category = $this->categoryService->delete($category_id);
 
             return response()->success(
@@ -206,42 +131,18 @@ class CategoryController extends Controller
         } catch (\Throwable $th) {
             return response()->error(
                 $th->getMessage(),
-                Response::HTTP_OK
+                400
             );
         }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
-     */
-    public function forceDelete()
+    public function getSubDataForCategory(Request $request)//si
     {
         try {
-            $category_id = request('category_id');
-            $category = $this->categoryService->forceDelete($category_id);
-
-            return response()->success(
-                [
-                    'message' => 'Cateegory Deleted Successfully',
-                    'data' => $category
-                ],
-                Response::HTTP_OK
-            );
-        } catch (\Throwable $th) {
-            return response()->error(
-                $th->getMessage(),
-                Response::HTTP_OK
-            );
-        }
-    }
-
-    public function getSubDataForCategory()
-    {
-        try {
-            $category_id = request('category_id');
+            $validatedData = $request->validate([
+                'category_id' => ['required', 'integer', 'exists:categories,id'],
+            ]);
+            $category_id = $validatedData['category_id'];            
             $SubCategories = $this->categoryService->getSubDataForCategory($category_id);
             return response()->success(
                 $SubCategories
@@ -254,10 +155,13 @@ class CategoryController extends Controller
         }
     }
 
-    public function getSubForCategory()
+    public function getSubForCategory(Request $request)//si
     {
         try {
-            $category_id = request('category_id');
+            $validatedData = $request->validate([
+                'category_id' => ['required', 'integer', 'exists:categories,id'],
+            ]);
+            $category_id = $validatedData['category_id'];   
             $SubCategories = $this->categoryService->getSubForCategory($category_id);
             return response()->success(
                 $SubCategories
@@ -269,14 +173,49 @@ class CategoryController extends Controller
             );
         }
     }
-    
-    
-    public function getImageUrl(){
-        
-        $publicId = "photo/(7)";
-        
-      return  $this->categoryService->getcImageUrl($publicId);
-        
+
+    public function changeVisibility(Request $request){
+        try{
+			$validate = Validator::make(
+				$request->all(),
+                [
+					'category_id' => 'required|exists:categories,id',					
+                ]
+            );
+
+            if ($validate->fails()) {
+                return response()->error(
+                    $validate->errors()
+                    ,Response::HTTP_UNPROCESSABLE_ENTITY
+                );
+            }
+			
+			$category_id = request('category_id');
+			$category = Category::findOrFail($category_id);
+			$is_valid = $category->valid;
+			if($is_valid){
+                $category->update(['valid'=>0]);
+                $sub_category_ids = $category->subCategories->pluck('id');
+                SubCategory::whereIn('id',$sub_category_ids)->update(['valid' => 0]);
+                $products = Product::whereIn('sub_category_id', $sub_category_ids)->get();
+                Product::whereIn('id', $products->pluck('id'))->update(['available' => 0]);	
+			}
+			else{
+                $category->update(['valid'=>1]);
+                $sub_category_ids = $category->subCategories->pluck('id');
+                SubCategory::whereIn('id',$sub_category_ids)->update(['valid' => 1]);
+                $products = Product::whereIn('sub_category_id', $sub_category_ids)->get();
+                Product::whereIn('id', $products->pluck('id'))->update(['available' => 1]);	
+			}
+			return response()->success(
+                'Category status was updated, this include all its subcategories and products'
+            , Response::HTTP_OK);
+        	
+		}catch(InvalidArgumentException $e) {
+            return response()->error(
+                $e->getMessage()
+            , Response::HTTP_NOT_FOUND);
+        }
     }
     
 }
